requirements.txt
absl-py==2.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
banks==2.1.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
biopython==1.85
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
coredis==4.22.0
cryptography==45.0.4
cssselect2==0.8.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dm-tree==0.1.9
docopt==0.6.2
docstring_parser==0.16
ecdsa==0.19.1
emoji==2.14.1
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
feedparser==6.0.10
fhaviary==0.19.0
fhlmi==0.27.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
fonttools==4.58.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
futurehouse-client==0.3.19
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-aiplatform==1.97.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.19.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gpt-researcher==0.13.3
graphdatascience==1.15.1
greenlet==3.2.3
griffe==1.7.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
guidance==0.2.1
guidance-stitch==0.1.5
h11==0.16.0
hf-xet==1.1.3
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.6
idna==2.8
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
json_repair==0.46.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kombu==5.5.4
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.22
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.45
ldp==0.27.0
limits==5.2.0
litellm==1.67.4.post1
llama-cloud==0.1.26
llama-cloud-services==0.6.34
llama-index==0.12.43
llama-index-agent-openai==0.4.11
llama-index-cli==0.4.3
llama-index-core==0.12.43
llama-index-embeddings-huggingface==0.5.4
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.7
llama-index-instrumentation==0.1.0
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-index-workflows==0.2.2
llama-parse==0.6.34
llguidance==0.6.31
loguru==0.7.3
lxml==5.4.0
lxml_html_clean==0.4.2
Markdown==3.8
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
md2pdf==1.0.1
mistune==3.1.3
mpmath==1.3.0
multidict==6.4.4
multimethod==2.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neo4j==5.28.1
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.2.6
olefile==0.47
openai==1.86.0
openreview==0.0.2
openreview-py==1.49.1
ordered-set==4.1.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pdfkit==1.0.0
pdfminer.six==20250506
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
powerlaw==1.5
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
Pympler==1.1
PyMuPDF==1.26.0
pyparsing==3.2.3
pypdf==5.6.0
PyPDF2==3.0.1
pyphen==0.17.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-jose==3.5.0
python-json-logger==3.3.0
python-louvain==0.16
python-magic==0.4.27
python-multipart==0.0.6
python-oxmsg==0.0.2
python-slugify==8.0.4
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==27.0.0
RapidFuzz==3.13.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==4.1.0
sgmllib3k==1.0.0
shapely==2.1.1
simsimd==6.4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
textdistance==4.6.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tld==0.13.1
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.36.0
uri-template==1.3.0
urllib3==1.25.11
usearch==2.17.10
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
yachalk==0.1.7
yarl==1.20.1
zipp==3.23.0
zopfli==0.2.3.post1
zstandard==0.23.0
